src/requirements.txt
Django==2.1.1
djangorestframework==3.8.2
requests==2.19.1
transmissionrpc==0.11
tmdbsimple==2.2.0
jsonfield2==3.0.1
gunicorn==19.9.0
regex==2018.8.29
bpython==0.17.1
redis==2.10.6
whitenoise==4.1
dj-database-url==0.5.0
celery==4.2.1
redis==2.10.6
flower==0.9.2
